Add cluster manager support.
Separate out the node subnet allocation and hybrid node
subnet allocation from ovnkube-master to a new mode -
cluster-manager.  New mode can be enabled by
passing "--init-cluster-manager <NODE_NAME>" to "ovnkube".
Even though it is technically possible to run ovnkube with
both "--init-cluster-manager" and "--init-master" mode, it
is recommended to run them separately.  The kind script is
enhanced to deploy cluster-manager in a separate
deployment - ovnkube-cluster-manager.

When deployments update/upgrade ovn-kubernetes to a version
which has this support, they should make sure to run cluster-manager
as a separate container in the existing ovnkube-master deployment
or as a separate deployment.

Signed-off-by: Numan Siddique <numans@ovn.org>